requirements.txt
requests~=2.27.1
openai~=0.16.0
gtts~=2.2.4
sox~=1.4.1
pyairtable~=1.1.0
nltk~=3.7
pandas~=1.4.1
deep_translator~=1.8.3
pyTelegramBotAPI~=4.4.0
internetarchive
moviepy
google-cloud-texttospeech
tqdm
srt
ffmpeg-python
